Move the migration code away from db_storage.rs
