+++ linera-storage/src/lib.rs
@@ -4,6 +4,7 @@
 //! This module defines the storage abstractions for individual chains and certificates.
 
 mod db_storage;
+mod migration;
 
 use std::{collections::BTreeMap, ops::RangeInclusive, sync::Arc};
